# docs/swagger.yaml
basePath: /v1
definitions:
  models.Translation:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
      translation:
        example: text for translation
        type: string
    type: object
  v1.doTranslateRequest:
    properties:
      destination:
        example: en
        type: string
      original:
        example: текст для перевода
        type: string
      source:
        example: auto
        type: string
    required:
    - destination
    - original
    - source
    type: object
  v1.historyResponse:
    properties:
      history:
        items:
          $ref: '#/definitions/models.Translation'
        type: array
    type: object
  v1.response:
    properties:
      error:
        example: message
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Using a translation service as an example
  title: Go Clean Template API
  version: "1.0"
paths:
  /translation/do-translate:
    post:
      consumes:
      - application/json
      description: Translate a text
      operationId: do-translate
      parameters:
      - description: Set up translation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/v1.doTranslateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Translation'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Translate
      tags:
      - translation
  /translation/history:
    get:
      consumes:
      - application/json
      description: Show all translation history
      operationId: history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.historyResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Show history
      tags:
      - translation
swagger: "2.0"